move _input_manipulations to utils, rename _disparities to _fairness
