+++ fairlearn/metrics/__init__.py
@@ -32,9 +32,9 @@ from ._base_metrics import mean_prediction  # noqa: F401
 from ._base_metrics import selection_rate  # noqa: F401
 from ._base_metrics import true_negative_rate  # noqa: F401
 from ._base_metrics import true_positive_rate  # noqa: F401; noqa: F401
-from ._disparities import demographic_parity_difference  # noqa: F401
-from ._disparities import demographic_parity_ratio  # noqa: F401
-from ._disparities import equalized_odds_difference, equalized_odds_ratio  # noqa: F401
+from ._fairness import demographic_parity_difference  # noqa: F401
+from ._fairness import demographic_parity_ratio  # noqa: F401
+from ._fairness import equalized_odds_difference, equalized_odds_ratio  # noqa: F401
 from ._generated_metrics import _generated_metric_dict  # noqa: F401
 from ._make_derived_metric import make_derived_metric  # noqa: F401
 from ._metric_frame import MetricFrame  # noqa: F401
@@ -55,7 +55,7 @@ _core = [
     "plot_model_comparison"
 ]
 
-_disparities = [
+_fairness = [
     "demographic_parity_difference",
     "demographic_parity_ratio",
     "equalized_odds_difference",
@@ -73,5 +73,5 @@ _base_metrics = [
 ]
 
 __all__ = (
-    _core + _disparities + _base_metrics + list(sorted(_generated_metric_dict.keys()))
+    _core + _fairness + _base_metrics + list(sorted(_generated_metric_dict.keys()))
 )
