+++ fairlearn/metrics/_base_metrics.py
@@ -13,7 +13,7 @@ from typing import Any
 from ._mean_predictions import mean_prediction  # noqa: F401
 from ._metric_frame import check_consistent_length
 from ._selection_rate import selection_rate  # noqa: F401,E501
-from ._input_manipulations import _convert_to_ndarray_and_squeeze
+from fairlearn.utils._input_manipulations import _convert_to_ndarray_and_squeeze
 
 _EMPTY_INPUT_PREDICTIONS_ERROR_MESSAGE = (
     "Empty y_pred passed to selection_rate function."
